+++ packages/host/vaster_cli/lib/src/vaster_cli_runner.dart
@@ -10,6 +10,7 @@ import 'commands/doctor_command.dart';
 import 'commands/inspect_command.dart';
 import 'commands/check_command.dart';
 import 'commands/eval_command.dart';
+import 'commands/replay_command.dart';
 import 'commands/resume_command.dart';
 import 'commands/run_command.dart';
 import 'commands/serve_command.dart';
@@ -27,6 +28,7 @@ class VasterCliRunner {
           CheckCommand(),
           EvalCommand(),
           ResumeCommand(),
+          ReplayCommand(),
           CompileCommand(),
           AuditCommand(),
         DebugCommand(),